+++ Events/ItemManager.cs
@@ -1,5 +1,4 @@
 ﻿using RoR2;
-using System.Reflection;
 using UnityEngine;
 using UnityEngine.Networking;