+++ UI/VoteItems.cs
@@ -1,10 +1,10 @@
-﻿using UnityEngine;
+﻿using RoR2;
 using RoR2.UI;
 using System;
-using RoR2;
-using System.Reflection;
-using UnityEngine.UI;
 using System.Collections.Generic;
+using System.Reflection;
+using UnityEngine;
+using UnityEngine.UI;
 
 namespace VsTwitch
 {